refactor: rename const
