+++ src/App.tsx
@@ -1,7 +1,7 @@
 import Footer from 'components/Footer/Footer';
 import Header from 'components/Header/Header';
 import React from 'react';
-import { routes } from 'utils/const/routes';
+import { ROUTES } from 'utils/const/routes';
 import { Routes, Route, Navigate } from 'react-router-dom';
 import './App.css';
 import Page404 from 'pages/Page404/Page404';
@@ -14,11 +14,11 @@ function App() {
     <>
       <Header />
       <Routes>
-        <Route path={routes.homePage} element={<Home />} />
-        <Route path={routes.signInPage} element={<SignIn />} />
-        <Route path={routes.signUpPage} element={<SignUp />} />
-        <Route path={routes.notFoundPage} element={<Page404 />} />
-        <Route path={routes.anyPage} element={<Navigate to={routes.notFoundPage} replace />} />
+        <Route path={ROUTES.HOME_PAGE} element={<Home />} />
+        <Route path={ROUTES.SIGN_IN_PAGE} element={<SignIn />} />
+        <Route path={ROUTES.SIGN_UP_PAGE} element={<SignUp />} />
+        <Route path={ROUTES.NOT_FOUND_PAGE} element={<Page404 />} />
+        <Route path={ROUTES.ANY_PAGE} element={<Navigate to={ROUTES.NOT_FOUND_PAGE} replace />} />
       </Routes>
       <Footer />
     </>
