+++ src/utils/const/host.ts
@@ -1,1 +1,1 @@
-export const host = 'https://final-task-backend-production-a3a6.up.railway.app';
+export const HOST = 'https://final-task-backend-production-a3a6.up.railway.app';
