+++ src/utils/const/routes.ts
@@ -1,7 +1,7 @@
-export const routes = {
-  homePage: '/',
-  signUpPage: '/signUp',
-  signInPage: '/signIn',
-  notFoundPage: '/notFoundPage',
-  anyPage: '*',
+export const ROUTES = {
+  HOME_PAGE: '/',
+  SIGN_UP_PAGE: '/signUp',
+  SIGN_IN_PAGE: '/signIn',
+  NOT_FOUND_PAGE: '/notFoundPage',
+  ANY_PAGE: '*',
 };
